Hides the application if focus is lost.
The application is hidden it is loses focus, as it cannot be activated
programatically for instance on Windows if another application is
activated. The application will simply flash in the taskbar instead.
Another option would be hiding and showing the window, but this leads
to unpleasent flickering, so the application is simply hidden.