Restricts to single application instance.
As running multiple instances causes a whole bunch of issues (hotkey
not working, state events not properly delivered and alike) there is
now a restriction that the application can be executed just once. If
the application is started a second time it will be terminated. The
single instance is enforced by placing an exclusive file lock on the
executable file. If it can be granted the application is assumed to
be the first instance and it is allowed; if the lock cannot be placed
the application is assumed to be an additional instance.