Adds support for configuration import.
The <import> tag can now be used to import other configuration, which
is then treated as part of the root configuration. Currently importing
files using the <file> element is supported. The imported file will
also be watched and therefore the model is automatically updated when
an imported file has changed. An import may for instance look like

   <items>
      <import>
         <file>second.xml</file>
         <file>third.xml</file>
      </import>
   </items>

A file that is imported twice or more often will be ignored.