Adds edit item option to the context menu.
The context menu of the item edit has now an action to edit the root
configuration and the context menu of the item view has an action to
edit the configuration of the respective item (which may also be an
imported configuration). As editor the default system editor is used.