Implements advanced search expression.
The search expression is no longer a regular expression, but rather a
simplified, but logical expression aware search expression. The '?'
and '*' wildcards can be used and the "AND" (default), "OR" and "NOT"
keywords can be used to form a logical expression. In addition, a tag
can be matched using the "@<tag>" notation (by default the name is
matched).