Implements on-demand window sizing.
Currently the window size is always fixed and the non-used part of the
result list was made invisible using a transparent background, which
basically prevented decoration of the result list and made the
transparent region non-clickable for other applications. The change
implements a mechanism that resizes the search window properly based on
the widgets sizes.