Changes default search to "match any".
Currently each search term is matching the start of an item only as it is
implcitily prepended with a caret. With this change the caret is no longer
prepended and thus the term matches anywhere in the item. If the previous
behaviour is desired the caret can be added explicitily to the search
term on search.